% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvts.R
\name{cvts}
\alias{cvts}
\title{Cross-Validation subsamples for time series}
\usage{
cvts(data, cv_size, lead_time, ...)
}
\arguments{
\item{data}{a dataframe with the time series data that the user wants to iterate.}

\item{cv_size}{Numeric. Cross-validation window that the user want to test the performance of a model.}

\item{lead_time}{Numeric. How many periods ahead (lead time) the user wants to predict.}

\item{...}{Other arguments associated with traints and testts functions}
}
\value{
a tibble with rows=iterations.
\itemize{
 \item iteration
 \item rest of the columns
}
}
\description{
This function takes traints and testts and create a single tibble with both results
}
\examples{
cvts(rainfall_cr, cv_size=6, lead_time=3, keep_prev=TRUE)
}
