% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_series.r
\name{nest_series}
\alias{nest_series}
\title{Clean and nest grouped series}
\usage{
nest_series(data, key, value, date)
}
\arguments{
\item{data}{a dataframe consisting in hierarchical series}

\item{key}{name of the column (unquoted) that will be used as a key}

\item{value}{column name of the univariate time series value}

\item{date}{column name of the date}
}
\value{
a nested series grouped by key
\itemize{
 \item cleaned series
}
}
\description{
Executes a set of sanity checks to prepare the analysis of monthly data
}
\examples{
nest_series(data=rainfall, key=country, value=rainfall_mm, date=date)
}
