% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Rainfall data}
\format{Tibble with 3 columns
\describe{
 \item{country}{Label of the country}
 \item{date}{Date in which the value was recorded}
 \item{rainfall_mm}{Numeric showing the amount of rainfall}
}}
\source{
https://climateknowledgeportal.worldbank.org/download-data
}
\usage{
rainfall
}
\description{
Dataset containing monthly rainfall data from 1991 to 2016
}
\keyword{datasets}
