% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testts.R
\name{testts}
\alias{testts}
\title{Cross-validated time series test set}
\usage{
testts(data, cv_size, lead_time, keep_prev = FALSE)
}
\arguments{
\item{data}{a dataframe with the time series data that the user wants to iterate.}

\item{cv_size}{Numeric. Cross-validation window that the user want to test the performance of a model.}

\item{lead_time}{Numeric. How many periods ahead (lead time) the user wants to predict.}

\item{keep_prev}{Logical. Controls if the test set includes previous data when \eqn{lead_time>2}}
}
\value{
a tibble with rows=iterations.
\itemize{
 \item iteration
 \item rest of the columns
}
}
\description{
Creates a cross-validated test set for a given number of time series iterations.
}
\details{
This function follows a tidy approach to create test sets applicable to time series data.
It admits tsibble, tibble and data frame objects.
}
\examples{
testts(rainfall, cv_size=6, lead_time=3, keep_prev=TRUE)
}
