% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traints.R
\name{traints}
\alias{traints}
\title{Cross-validated time series train set}
\usage{
traints(data, cv_size, lead_time)
}
\arguments{
\item{data}{a dataframe with the time series data that the user wants to iterate.}

\item{cv_size}{Numeric. Cross-validation window that the user want to test the performance of a model.}

\item{lead_time}{Numeric. How many periods ahead (lead time) the user wants to predict.}
}
\value{
a tibble with rows=iterations.
\itemize{
 \item iteration
 \item rest of the columns
}
}
\description{
Creates a cross-validated train set for a given number of time series iterations.
}
\details{
This function follows a tidy approach to create train sets applicable to time series data.
It admits tsibble, tibble and data frame objects.
}
\examples{
traints(rainfall_cr, cv_size=6, lead_time=3)
}
